[dev_cmds] Split integration test harness and test case definitions into separate files
