+++ dev_cmds/integration-tests/integration-test-group-factories.ts
@@ -1,0 +1,7 @@
+import { createMultiplePackageJsonsExampleTestGroup } from "./multiple-package-jsons-example";
+import { createSinglePackageJsonExampleTestGroup } from "./single-package-json-example";
+
+export const integrationTestGroupFactories = [
+  createSinglePackageJsonExampleTestGroup,
+  createMultiplePackageJsonsExampleTestGroup,
+];
