+++ dev_cmds/integration-tests/integration-test-harness/constants.ts
@@ -1,0 +1,3 @@
+export const VERDACCIO_CONTAINER_NAME = "devcmd_verdaccio";
+export const VERDACCIO_STORAGE_VOLUME_NAME = "devcmd_verdaccio_storage";
+export const LOCAL_REGISTRY_URL = "http://0.0.0.0:4873";
