+++ dev_cmds/utils/delay.ts
@@ -1,0 +1,7 @@
+export function delay(ms: number): Promise<void> {
+  return new Promise((resolve) => {
+    setTimeout(() => {
+      resolve();
+    }, ms);
+  });
+}
