+++ dev_cmds/utils/fs-utils.ts
@@ -1,0 +1,11 @@
+import fs from "fs-extra";
+
+export async function isFile(path: string): Promise<boolean> {
+  try {
+    const info = await fs.stat(path);
+    return info.isFile();
+  } catch (error) {
+    if (error.code === "ENOENT") return false; // TODO double-check code and comparison value
+    throw error;
+  }
+}
